+++ example/testfloat_compress.c
@@ -36,7 +36,7 @@ void cost_end()
 
 int main(int argc, char * argv[])
 {
-    int r5=0,r4=0,r3=0,r2=0,r1=0;
+    size_t r5=0,r4=0,r3=0,r2=0,r1=0;
     char outDir[640], oriFilePath[640], outputFilePath[640];
     char *cfgFile;
     
@@ -66,7 +66,7 @@ int main(int argc, char * argv[])
 	exit(0);
     sprintf(outputFilePath, "%s.sz", oriFilePath);
    
-    int nbEle;
+    size_t nbEle;
     float *data = readFloatData(oriFilePath, &nbEle, &status);
     if(status != SZ_SCES)
     {
@@ -76,7 +76,7 @@ int main(int argc, char * argv[])
     //float *revValue = (float *)malloc(sizeof(float));
     //*revValue = 1.0E36;
    
-    int outSize; 
+    size_t outSize; 
     //char *bytes = (char *)malloc(nbEle*sizeof(float)); //
     //SZ_compress_args2(SZ_FLOAT, data, bytes, &outSize, ABS, 0.0001, 0.0001, r5, r4, r3, r2, r1);    
     //char *bytes = SZ_compress_rev(SZ_FLOAT, data, revValue, &outSize, r5, r4, r3, r2, r1);
